Struct awscreds::Credentials

pub struct Credentials {
    pub access_key: Option<String>,
    pub secret_key: Option<String>,
    pub security_token: Option<String>,
    pub session_token: Option<String>,
}

AWS access credentials: access key, secret key, and optional token.

Example

Loads from the standard AWS credentials file with the given profile name, defaults to "default".

use awscreds::Credentials;

// Load credentials from `[default]` profile
let credentials = Credentials::default();

// Also loads credentials from `[default]` profile
let credentials = Credentials::new(None, None, None, None, None);

// Load credentials from `[my-profile]` profile
let credentials = Credentials::new(None, None, None, None, Some("my-profile".into()));

// Use anonymous credentials for public objects let credentials = Credentials::anonymous();

Credentials may also be initialized directly or by the following environment variables:

AWS_ACCESS_KEY_ID,

AWS_SECRET_ACCESS_KEY

AWS_SESSION_TOKEN

The order of preference is arguments, then environment, and finally AWS credentials file.

use awscreds::Credentials;

// Load credentials directly
let access_key = "AKIAIOSFODNN7EXAMPLE";
let secret_key = "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY";
let credentials = Credentials::new(Some(access_key), Some(secret_key), None, None, None);

// Load credentials from the environment
use std::env;
env::set_var("AWS_ACCESS_KEY_ID", "AKIAIOSFODNN7EXAMPLE");
env::set_var("AWS_SECRET_ACCESS_KEY", "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY");
let credentials = Credentials::new(None, None, None, None, None);

Fields
access_key: Option<String>

AWS public access key.

secret_key: Option<String>

AWS secret key.

security_token: Option<String>

Temporary token issued by AWS service.

session_token: Option<String>
Implementations

impl Credentials

pub fn from_sts_env(session_name: &str) -> Result<Credentials>

pub fn from_sts(
role_arn: &str,
session_name: &str,
web_identity_token: &str
) -> Result<Credentials>

pub fn default() -> Result<Credentials>

pub fn anonymous() -> Result<Credentials>

pub fn new(
access_key: Option<&str>,
secret_key: Option<&str>,
security_token: Option<&str>,
session_token: Option<&str>,
profile: Option<&str>
) -> Result<Credentials>

Initialize Credentials directly with key ID, secret key, and optional token.

pub fn from_env_specific(
access_key_var: Option<&str>,
secret_key_var: Option<&str>,
security_token_var: Option<&str>,
session_token_var: Option<&str>
) -> Result<Credentials>

pub fn from_env() -> Result<Credentials>

pub fn from_instance_metadata() -> Result<Credentials>

pub fn from_profile(section: Option<&str>) -> Result<Credentials>